feat(Ghosty): add idle variant and refactor character components
- Introduced GhostyIdle component for a non-waving variant of the Ghosty character.
- Refactored Ghosty component to utilize a shared GhostyCharacter for consistent rendering.
- Added shared SVG definitions and paths for better maintainability.
- Updated RemotionRoot to include the new GhostyIdle composition.
