+++ remotion/src/Ghosty/lib/index.ts
@@ -1,0 +1,4 @@
+export { GhostyCharacter, ghostySchema } from "./GhostyCharacter";
+export type { GhostyProps } from "./GhostyCharacter";
+export { GhostyDefs } from "./Defs";
+export * from "./paths";
