+++ remotion/src/Ghosty/lib/paths.ts
@@ -1,0 +1,29 @@
+// SVG path constants for the Ghosty character. ViewBox is 1000x1000.
+
+export const BODY_PATH =
+  "M500,240 " +
+  "C700,240 855,420 860,625 " +
+  "C863,770 810,860 700,895 " +
+  "Q500,930 300,895 " +
+  "C190,860 137,770 140,625 " +
+  "C145,420 300,240 500,240 Z";
+
+export const LEFT_LEG_PATH =
+  "M310,880 " +
+  "C220,895 150,940 175,985 " +
+  "C200,1020 290,1020 360,990 " +
+  "C420,965 430,915 400,890 Z";
+
+export const RIGHT_LEG_PATH =
+  "M690,880 " +
+  "C780,895 850,940 825,985 " +
+  "C800,1020 710,1020 640,990 " +
+  "C580,965 570,915 600,890 Z";
+
+// Raised mitten arm sticking up-right from the right shoulder.
+export const ARM_PATH =
+  "M800,560 " +
+  "C870,485 955,500 955,575 " +
+  "C955,635 880,650 800,615 Z";
+
+export const VIEWBOX = 1000;
